# backend/internal/pkg/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HRIS Platform API",
...
    "version": "1.5.0",
    "contact": {
      "name": "HRIS Platform Team",
      "email": "dev@hris-platform.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development"
    }
  ],
...
  "paths": {
...
    "/api/v1/tenant/employee-movements/movements": {
      "get": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "List employee movements with pagination",
        "operationId": "listMovements",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated list of employee movements" } }
      },
      "post": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Create a new employee movement (promotion, demotion, mutation, etc.)",
        "operationId": "createMovement",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateMovementRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Employee movement created" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/v1/tenant/employee-movements/movements/{id}": {
      "get": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Get employee movement by ID",
        "operationId": "getMovement",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Employee movement details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Update employee movement (draft only)",
        "operationId": "updateMovement",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateMovementRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Employee movement updated" } }
      },
      "delete": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Delete employee movement (draft only)",
        "operationId": "deleteMovement",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Employee movement deleted" } }
      }
    },
    "/api/v1/tenant/employee-movements/movements/{id}/approve": {
      "post": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Approve an employee movement (changes status from draft to approved)",
        "operationId": "approveMovement",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Movement approved" }, "409": { "description": "Conflict: cannot approve (e.g., not in draft status)" } }
      }
    },
    "/api/v1/tenant/employee-movements/movements/{id}/execute": {
      "post": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Execute an approved employee movement (changes status from approved to executed)",
        "operationId": "executeMovement",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Movement executed" }, "409": { "description": "Conflict: cannot execute (e.g., not in approved status)" } }
      }
    },
    "/api/v1/tenant/employee-movements/movements/{id}/cancel": {
      "post": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Cancel an employee movement (draft or approved only)",
        "operationId": "cancelMovement",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Movement cancelled" }, "409": { "description": "Conflict: cannot cancel (e.g., already executed)" } }
      }
    },
    "/api/v1/tenant/employee-movements/employees/{employeeId}/movements": {
      "get": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "List movements by employee ID",
        "operationId": "listMovementsByEmployee",
        "parameters": [
          { "name": "employeeId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated list of movements for the employee" } }
      }
    },
    "/api/v1/tenant/employee-movements/contracts": {
      "get": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "List employee contracts with pagination",
        "operationId": "listContracts",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated list of employee contracts" } }
      },
      "post": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Create a new employee contract (PKWT/PKWTT/Daily)",
        "operationId": "createContract",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateContractRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Employee contract created" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/v1/tenant/employee-movements/contracts/{id}": {
      "get": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Get employee contract by ID",
        "operationId": "getContract",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Employee contract details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Update employee contract",
        "operationId": "updateContract",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateContractRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Employee contract updated" } }
      },
      "delete": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "Delete employee contract",
        "operationId": "deleteContract",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Employee contract deleted" } }
      }
    },
    "/api/v1/tenant/employee-movements/employees/{employeeId}/contracts": {
      "get": {
        "tags": ["Tenant: Employee Movement & Career Management"],
        "summary": "List contracts by employee ID",
        "operationId": "listContractsByEmployee",
        "parameters": [
          { "name": "employeeId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Paginated list of contracts for the employee" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Access Token. Dapatkan dengan login ke platform endpoint."
      }
    },
    "schemas": {
...
      "CreateMovementRequest": {
        "type": "object",
        "required": ["employee_id", "movement_type", "decision_letter_number", "decision_letter_date", "effective_date"],
        "properties": {
          "employee_id": { "type": "string", "format": "uuid", "description": "ID of the employee" },
          "movement_type": { "type": "string", "enum": ["promotion", "demotion", "mutation", "contract_extension", "status_change", "retirement", "offboarding", "other"], "example": "promotion" },
          "from_employment_id": { "type": "string", "format": "uuid", "nullable": true },
          "to_employment_id": { "type": "string", "format": "uuid", "nullable": true },
          "from_organization_id": { "type": "string", "format": "uuid", "nullable": true },
          "to_organization_id": { "type": "string", "format": "uuid", "nullable": true },
          "from_position_id": { "type": "string", "format": "uuid", "nullable": true },
          "to_position_id": { "type": "string", "format": "uuid", "nullable": true },
          "from_employment_status_id": { "type": "string", "format": "uuid", "nullable": true },
          "to_employment_status_id": { "type": "string", "format": "uuid", "nullable": true },
          "reason": { "type": "string", "nullable": true },
          "decision_letter_number": { "type": "string", "example": "SK-2026-001" },
          "decision_letter_date": { "type": "string", "format": "date", "example": "2026-07-01" },
          "effective_date": { "type": "string", "format": "date", "example": "2026-08-01" },
          "notes": { "type": "string", "nullable": true }
        }
      },
      "UpdateMovementRequest": {
        "type": "object",
        "properties": {
          "movement_type": { "type": "string", "enum": ["promotion", "demotion", "mutation", "contract_extension", "status_change", "retirement", "offboarding", "other"], "nullable": true },
          "to_organization_id": { "type": "string", "format": "uuid", "nullable": true },
          "to_position_id": { "type": "string", "format": "uuid", "nullable": true },
          "to_employment_status_id": { "type": "string", "format": "uuid", "nullable": true },
          "reason": { "type": "string", "nullable": true },
          "decision_letter_number": { "type": "string", "nullable": true },
          "decision_letter_date": { "type": "string", "format": "date", "nullable": true },
          "effective_date": { "type": "string", "format": "date", "nullable": true },
          "status": { "type": "string", "enum": ["draft", "approved", "executed", "cancelled"], "nullable": true },
          "notes": { "type": "string", "nullable": true }
        }
      },
...
      "CreateContractRequest": {
        "type": "object",
        "required": ["employee_id", "contract_number", "contract_type", "start_date"],
        "properties": {
          "employee_id": { "type": "string", "format": "uuid" },
          "contract_number": { "type": "string", "example": "CTR-2026-001" },
          "contract_type": { "type": "string", "enum": ["pkwt", "pkwtt", "daily", "other"], "example": "pkwt" },
          "start_date": { "type": "string", "format": "date", "example": "2026-01-01" },
          "end_date": { "type": "string", "format": "date", "nullable": true, "example": "2026-12-31" },
          "previous_contract_id": { "type": "string", "format": "uuid", "nullable": true },
          "decision_letter_number": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true },
          "document_url": { "type": "string", "format": "uri", "nullable": true }
        }
      },
      "UpdateContractRequest": {
        "type": "object",
        "properties": {
          "contract_number": { "type": "string", "nullable": true },
          "contract_type": { "type": "string", "enum": ["pkwt", "pkwtt", "daily", "other"], "nullable": true },
          "end_date": { "type": "string", "format": "date", "nullable": true },
          "decision_letter_number": { "type": "string", "nullable": true },
          "notes": { "type": "string", "nullable": true },
          "document_url": { "type": "string", "format": "uri", "nullable": true },
          "status": { "type": "string", "enum": ["active", "expired", "extended", "terminated"], "nullable": true }
        }
      },
...
    }
  }
}